fix(helpdesk_privacy): ensure that we do not subscribe users who do not already follow the ticket when processing a message_update. fix(brands): fix due to upstream changes under https://github.com/odoo/odoo/commit/09d5f83ee88c1f786a678d0aea6de85c2391505c
This is not a perfect implementation, but I'm unclear how to better handle better without a lot of "what-ifs":

i.e. In the scenario User A who is following a private ticket CC's User B who is not, Odoo needs to accept the email, but we cannot just hard block the email as it should be posted. But we also can't just blackhole it.

This then needs to handle BCC, etc.

What I've chosen to do is accept the email, post it, but ensure that only the existing users are subscribed.

This has the side effect of allowing parties without access the ability to potentially post message, but not be auto-subscribed.

Fixes GH181524